Support decimal scores for meets
BoysScore/BoysOpponentScore/GirlsScore/GirlsOpponentScore changed
from int? to decimal? in model, ViewModel, and TeamResultViewModel.
Score labels use 0.## format to strip trailing zeros (87.5 not 87.50).
Score inputs get step="any". DB columns widened to DECIMAL(8,2).

+++ CloverleafTrack.Models/Meet.cs
@@ -30,10 +30,10 @@ public class Meet
     public string? EntryNotes { get; set; } = string.Empty;
 
     // Optional team score / placement (entered post-meet)
-    public int? BoysScore { get; set; }
-    public int? BoysOpponentScore { get; set; }
-    public int? GirlsScore { get; set; }
-    public int? GirlsOpponentScore { get; set; }
+    public decimal? BoysScore { get; set; }
+    public decimal? BoysOpponentScore { get; set; }
+    public decimal? GirlsScore { get; set; }
+    public decimal? GirlsOpponentScore { get; set; }
     public int? BoysPlace { get; set; }
     public int? GirlsPlace { get; set; }
     public int? FieldSize { get; set; }
